feature: 代码生成

+++ user-center/src/main/java/com/ziyao/harbor/usercenter/dto/MenuDTO.java
@@ -18,7 +18,7 @@ import java.util.Objects;
  * </p>
  *
  * @author zhangziyao
- * @since 2023-05-09
+ * @since 2024-06-08
  */
 @Data
 public class MenuDTO implements EntityDTO<Menu>, Serializable {
@@ -57,7 +57,7 @@ public class MenuDTO implements EntityDTO<Menu>, Serializable {
     /**
      * 资源级别
      */
-    private Integer level;
+    private Byte level;
     /**
      * 排序
      */
